Refactored to masterStore
